Added comments & fixed code
I added comments to most lines that weren't straight forward.

changed the output pin #defines to const int16_t to foce those to conform to their types and added pin declarations for the joystick and button pins as well for clarity sake.

I'm using 2 separate linear pots to test the setup. Still a bit hazy on how the ArduinoJoystick lib works but I think I'm getting the hang of it.

constrained the joystick values to within -255 to 255. This may not be needed if the joystick is calibrated correctly but im just using 2 pots and didnt want to change the calibration settings for the joystick - plus it's a bit more insurance that we dont get an out of range value

changed the analogWrites on the valve pins to use the absolute value of the given axis value so we dont try to analog write a negative value.